changed table definition removed migration from bookshelf.js config file -> see readme on how to do it adjusted server.js: removed transformation for date locale to D/Y/M -> format of csv fits us locale format

+++ server/migrations/employees.js
@@ -16,7 +16,7 @@ exports.up = function(knex, Promise) {
         table.string('employment_status');
         table.string('department');
         table.string('position');
-        table.integer('pay_rate');
+        table.double('pay_rate');
         table.string('employee_source');
         table.string('performance_score');
         table.string('date_of_termination');
